+++ src/KeyForge.Core/Models/ThemeMode.cs
@@ -1,8 +1,0 @@
-namespace KeyForge.Core.Models;
-
-public enum ThemeMode
-{
-    Dark,
-    Light,
-    System
-}